drivers: qti_battery_charger: Align xm_property_id with stock
Checking qti_battery_charger.ko driver found in HyperOS can give
us some info about what's wrong with xm_property_id so we can align
it correctly. The result is that:
- Added 1 property with fully implemented code: XM_PROP_THERMAL_BOARD_TEMP.
The code is taken from chenfeng-u-oss and it include both show/store
functions and extra charger_notifier_event.
- Added 3 properties that have no code publicly available. These
require either Xiaomi to open the source or reverse engineer the driver including:
+ XM_PROP_NTC_ALARM
+ XM_PROP_BLANK_STATUS
+ XM_PROP_SCREEN_CCTOG
- Added some useless properties. For other devices it may serve the purpose of
checking dual-battery, checking pen battery or checking DP plugged in. These
are not available on socrates and looking at chenfeng-u-oss or sheng-u-oss we
can see that they are just filler for the enum.

Signed-off-by: hmtheboy154 <buingoc67@gmail.com>